// src/store/bookmarks/bookmarks.types.ts
export enum FETCH_BOOKMARKS_ACTION_TYPES {
   FETCH_BOOKMARKS_START = 'bookmarks/FETCH_BOOKMARKS_START',
   FETCH_BOOKMARKS_SUCCESS = 'bookmarks/FETCH_BOOKMARKS_SUCCESS',
   FETCH_BOOKMARKS_FAILED = 'bookmarks/FETCH_BOOKMARKS_FAILED',
   REFRESH_BOOKMARKS_START = 'bookmarks/REFRESH_BOOKMARKS_START',
   REFRESH_BOOKMARKS_SUCCESS = 'bookmarks/REFRESH_BOOKMARKS_SUCCESS',
   REFRESH_BOOKMARKS_FAILED = 'bookmarks/REFRESH_BOOKMARKS_FAILED',
}

export enum ADD_BOOKMARK_ACTION_TYPES {
   ADD_BOOKMARK_START = 'bookmarks/ADD_BOOKMARK_START',
   ADD_BOOKMARK_SUCCESS = 'bookmarks/ADD_BOOKMARK_SUCCESS',
   ADD_BOOKMARK_FAILED = 'bookmarks/ADD_BOOKMARK_FAILED',
}

export enum REMOVE_BOOKMARKS_ACTION_TYPES {
   REMOVE_BOOKMARK_START = 'bookmarks/REMOVE_BOOKMARK_START',
   REMOVE_BOOKMARK_SUCCESS = 'bookmarks/REMOVE_BOOKMARK_SUCCESS',
   REMOVE_BOOKMARK_FAILED = 'bookmarks/REMOVE_BOOKMARK_FAILED',
}